changed the dest files of my functions to avoid creating unnecessary new files.
